backend/src/requirements.txt
Django==3.1
djangorestframework_jwt==1.11.0
PyMySQL==0.10.0
djangorestframework==3.11.1
